Expose the vulkan function pointers so plugins don't need to load anything
